requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2
web3==6.11.1
openai==1.3.7
python-dotenv==1.0.0
ecdsa==0.18.0
ipfshttpclient==0.8.0
python-multipart==0.0.6
websockets==12.0
pytest==7.4.3
httpx==0.25.2